fix: avoid versioning override import cycle
