+++ packages/common/src/internal-workflow/index.ts
@@ -8,7 +8,6 @@ export { SYSTEM_NEXUS_PAYLOAD_METADATA_VALUE } from './system-nexus';
 export { SYSTEM_NEXUS_CONTEXT_METADATA_KEY } from './system-nexus';
 // ts-prune-ignore-next
 export { TEMPORAL_SYSTEM_NEXUS_ENDPOINT } from './system-nexus';
-export { versioningOverrideToProto } from './versioning-override';
 export {
   filterNullAndUndefined,
   mergeObjects,
