+++ packages/common/src/worker-deployments.ts
@@ -1,4 +1,4 @@
-import type { coresdk, temporal } from '@temporalio/proto';
+import { temporal, type coresdk } from '@temporalio/proto';
 import { makeProtoEnumConverters } from './internal-workflow';
 
 /**
@@ -49,6 +49,30 @@ export const [encodeVersioningBehavior, decodeVersioningBehavior] = makeProtoEnu
  * Represents versioning overrides. For example, when starting workflows.
  */
 export type VersioningOverride = PinnedVersioningOverride | 'AUTO_UPGRADE';
+
+/** @internal */
+export function versioningOverrideToProto(
+  versioningOverride: VersioningOverride | undefined
+): temporal.api.workflow.v1.IVersioningOverride | undefined {
+  if (versioningOverride == null) {
+    return undefined;
+  }
+  // TODO: Remove deprecated field assignments when versioning is non-experimental.
+  if (versioningOverride === 'AUTO_UPGRADE') {
+    return {
+      autoUpgrade: true,
+      behavior: temporal.api.enums.v1.VersioningBehavior.VERSIONING_BEHAVIOR_AUTO_UPGRADE,
+    };
+  }
+  return {
+    pinned: {
+      version: versioningOverride.pinnedTo,
+      behavior: temporal.api.workflow.v1.VersioningOverride.PinnedOverrideBehavior.PINNED_OVERRIDE_BEHAVIOR_PINNED,
+    },
+    behavior: temporal.api.enums.v1.VersioningBehavior.VERSIONING_BEHAVIOR_PINNED,
+    pinnedVersion: toCanonicalString(versioningOverride.pinnedTo),
+  };
+}
 
 /**
  * Workflow will be pinned to a specific deployment version.
